requirements.txt
Flask
aiohttp
requests
beautifulsoup4
python-whois
dnspython
